\name{coef.PICBayes}
\alias{coef.PICBayes}
\title{Coef method for a PICBayes model}
\description{Extracts estimated regression coefficients.}
\usage{
\method{coef}{PICBayes}(object, ...)}
\arguments{
  \item{object}{The class PICBayes object.}
  \item{\dots}{Other arguments if any.}
}
\value{An object of class \code{coef}.}