\name{logLik.PICBayes}
\alias{logLik.PICBayes}
\title{LogLik method for a PICBayes model}
\description{The log-likelihood of the observed partly 
interval-censored data estimated by log pseudo-marginal likelihood.}
\usage{
\method{logLik}{PICBayes}(object, ...)}
\arguments{
  \item{object}{Class PICBayes object.}
  \item{\dots}{Other arguments if any.}
}
\value{An object of class \code{logLik}.}
