\name{PICBayes-package}
\alias{PICBayes-package}
\docType{package}
\title{Bayesian Models for Partly Interval-Censored Data and General Interval-Censored Data}
\description{
Contains functions to fit proportional hazards (PH) model to partly interval-censored (PIC) data (Pan et al. (2020) <doi:10.1177/0962280220921552>), PH model with spatial frailty to spatially dependent PIC data (Pan and Cai (2021) <doi:10.1080/03610918.2020.1839497>), and mixed effects PH model to clustered PIC data. Each random intercept/random effect can follow both a normal prior and a Dirichlet process mixture prior. It also includes the corresponding functions for general interval-censored data.
}
\details{
\tabular{ll}{
Package: \tab PICBayes\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2021-08-04\cr
License: \tab GPL>=2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Chun Pan

Maintainer: Chun Pan \email{chunpan2003@hotmail.com}
}
