\name{summary.PICBayes}
\alias{summary.PICBayes}
\title{Summary method for a PICBayes model}
\description{Present output from function \code{\link{PICBayes}}.}
\usage{
\method{summary}{PICBayes}(object, ...)}
\arguments{
  \item{object}{Class PICBayes object.}
  \item{\dots}{Other arguments if any.}
}
\value{An object of class \code{summary}.}
